Convert standalone functions to class methods, fix linters errors
