+++ src/modules/itemClass.js
@@ -1,0 +1,7 @@
+export default class Item {
+  constructor() {
+    this.description = '';
+    this.completed = false;
+    this.index = 0;
+  }
+}
